pyproject.toml
[project]
name = "stockrhythm-monorepo"
version = "0.1.0"
requires-python = ">=3.12"

[tool.uv.workspace]
members = ["apps/backend", "apps/mock_exchange", "packages/*"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.110.0",
    "websockets>=12.0",
    "pyotp>=2.9.0",
    "stockrhythm-sdk",
    "numpy",
    "pytest-cov>=7.0.0",
    "stockrhythm-cli",
    "typer>=0.9.0",
]

[tool.uv.sources]
stockrhythm-sdk = { workspace = true }
stockrhythm-cli = { workspace = true }

[tool.pytest.ini_options]
pythonpath = [
    ".",
    "packages/stockrhythm-sdk/src",
    "packages/stockrhythm-cli/src"
]
asyncio_mode = "strict"
